FED-247: changed the type of QueryPlanCost to f64 (#5328)

+++ apollo-federation/src/query_plan/mod.rs
@@ -16,7 +16,7 @@ pub mod generate;
 pub mod query_planner;
 pub(crate) mod query_planning_traversal;
 
-pub type QueryPlanCost = i64;
+pub type QueryPlanCost = f64;
 
 #[derive(Debug, Default)]
 pub struct QueryPlan {
